Replace replication implementation with replicon (#1361)
Authored-by: Charles Bournhonesque <cbournhonesque@tower-research.com>

+++ lightyear_avian/src/lag_compensation/history.rs
@@ -187,7 +187,7 @@ fn update_collider_history(
 
             // step 1. update the history buffer of the parent
             history.add_update(tick, (*parent_position, *parent_rotation, *parent_aabb));
-            history.clear_until_tick(tick - (config.max_collider_history_ticks as u16));
+            history.clear_until_tick(tick - (config.max_collider_history_ticks as u32));
 
             // step 2. update the child's Position, Rotation, Collider so that the avian spatial query
             //  can use the collider's aabb envelope for broad-phase collision detection
